Split node.py into multiple files
